refactor(tk): restructure module
