+++ token/src/main/kotlin/io/github/wliamp/tk/SecretHandler.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.token.data
+package io.github.wliamp.tk
 
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
@@ -7,18 +7,38 @@ import com.nimbusds.jose.jwk.JWKSet
 import com.nimbusds.jose.jwk.RSAKey
 import com.nimbusds.jose.jwk.source.JWKSource
 import com.nimbusds.jose.proc.SecurityContext
-import io.github.wliamp.token.config.TokenProperties
 import org.springframework.scheduling.annotation.EnableScheduling
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
-import java.util.Collections.emptySet
+import java.util.Collections
 import java.util.concurrent.atomic.AtomicReference
+import kotlin.collections.map
+
+interface SecretLoader {
+    fun loadPrivateJwksJson(): String
+}
+
+class EnvSecretLoader(private val props: Properties) : SecretLoader {
+    override fun loadPrivateJwksJson(): String =
+        System.getenv(props.envVar) ?: error("Env var ${props.envVar} not found")
+}
+
+data class PrivateKeySet(
+    val currentKid: String,
+    val graceKids: Set<String>,
+    val privateKeys: List<RSAKey>
+) {
+    fun active(): RSAKey = privateKeys.first { it.keyID == currentKid }
+    fun verificationPublicKeys(): List<RSAKey> =
+        privateKeys.filter { it.keyID == currentKid || it.keyID in graceKids }
+            .map { it.toPublicJWK().toRSAKey() }
+}
 
 @Component
 @EnableScheduling
 class KeySetManager(
     private val loader: SecretLoader,
-    private val props: TokenProperties,
+    private val props: Properties,
     private val objectMapper: ObjectMapper
 ) {
     private val cache = AtomicReference<PrivateKeySet>()
@@ -32,7 +52,7 @@ class KeySetManager(
         val json = loader.loadPrivateJwksJson()
         val root = objectMapper.readTree(json)
         val currentKid = root["currentKid"].asText()
-        val graceKids = root["graceKids"]?.map { it.asText() }?.toSet() ?: emptySet()
+        val graceKids = root["graceKids"]?.map { it.asText() }?.toSet() ?: Collections.emptySet()
         val keysNode = root["keys"]
         val privateKeys = keysNode.map { k ->
             val map: Map<String, Any> = objectMapper.convertValue(k, object : TypeReference<Map<String, Any>>() {})
