+++ token/src/main/kotlin/io/github/wliamp/tk/Token.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.token.data
+package io.github.wliamp.tk
 
 import java.time.Instant
 
@@ -9,3 +9,9 @@ data class Token(
     val expiration: Instant,
     val claims: Map<String, Any>
 )
+
+enum class Type {
+    ACCESS,
+    REFRESH,
+    SERVICE
+}
