--- token/src/main/kotlin/io/github/wliamp/token/data/SecretLoader.kt
@@ -1,5 +1,0 @@
-package io.github.wliamp.token.data
-
-interface SecretLoader {
-    fun loadPrivateJwksJson(): String
-}
