--- token/src/main/kotlin/io/github/wliamp/token/data/Type.kt
@@ -1,7 +1,0 @@
-package io.github.wliamp.token.data
-
-enum class Type {
-    ACCESS,
-    REFRESH,
-    SERVICE
-}
