+++ token/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.wliamp.token.config.TokenAutoConfig
+io.github.wliamp.tk.AutoConfig
