+++ token/src/test/kotlin/io/github/wliamp/tk/TestClass.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.token
+package io.github.wliamp.tk
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
